phase-2/backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.0
sqlmodel==0.0.16
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
